refactor: move DATADOG_STORAGE_TIER tests to separate file
Move the DATADOG_STORAGE_TIER environment variable tests from
logs.test.ts to a dedicated logs-storage-tier.test.ts file to
prevent MSW server conflicts during concurrent test execution.
