+++ tests/tools/logs.test.ts
@@ -1,5 +1,5 @@
 import { v2 } from '@datadog/datadog-api-client'
-import { describe, it, expect, beforeEach, afterEach } from 'vitest'
+import { describe, it, expect } from 'vitest'
 import { createDatadogConfig } from '../../src/utils/datadog'
 import { createLogsToolHandlers } from '../../src/tools/logs/tool'
 import { createMockToolRequest } from '../helpers/mock'
@@ -375,240 +375,6 @@ describe('Logs Tool', () => {
 
         expect(response.content[0].text).toContain('Services')
         expect(response.content[0].text).toContain('[]') // Empty array of services
-      })()
-
-      server.close()
-    })
-  })
-
-  describe('DATADOG_STORAGE_TIER environment variable', () => {
-    let originalEnv: string | undefined
-
-    beforeEach(() => {
-      // Save original env value
-      originalEnv = process.env.DATADOG_STORAGE_TIER
-    })
-
-    afterEach(() => {
-      // Restore original env value
-      if (originalEnv === undefined) {
-        delete process.env.DATADOG_STORAGE_TIER
-      } else {
-        process.env.DATADOG_STORAGE_TIER = originalEnv
-      }
-    })
-
-    it('should include storageTier in get_logs when DATADOG_STORAGE_TIER is set to flex', async () => {
-      process.env.DATADOG_STORAGE_TIER = 'flex'
-
-      let capturedRequestBody: { filter?: { storage_tier?: string } } | null =
-        null
-      const mockHandler = http.post(logsEndpoint, async ({ request }) => {
-        capturedRequestBody = await request.json()
-        return HttpResponse.json({
-          data: [
-            {
-              id: 'test-id',
-              attributes: {
-                timestamp: 1640995199999,
-                status: 'info',
-                message: 'Test log from flex tier',
-                service: 'test-service',
-              },
-              type: 'log',
-            },
-          ],
-          meta: { page: {} },
-        })
-      })
-
-      const server = setupServer(mockHandler)
-
-      await server.boundary(async () => {
-        const request = createMockToolRequest('get_logs', {
-          query: 'service:test-service',
-          from: 1640995100,
-          to: 1640995200,
-          limit: 10,
-        })
-        await toolHandlers.get_logs(request)
-
-        // Verify storage_tier was included in the filter (API client serializes storageTier as storage_tier)
-        expect(capturedRequestBody).toBeDefined()
-        expect(capturedRequestBody.filter).toBeDefined()
-        expect(capturedRequestBody.filter.storage_tier).toBe('flex')
-      })()
-
-      server.close()
-    })
-
-    it('should include storageTier in get_all_services when DATADOG_STORAGE_TIER is set to indexes', async () => {
-      process.env.DATADOG_STORAGE_TIER = 'indexes'
-
-      let capturedRequestBody: { filter?: { storage_tier?: string } } | null =
-        null
-      const mockHandler = http.post(logsEndpoint, async ({ request }) => {
-        capturedRequestBody = await request.json()
-        return HttpResponse.json({
-          data: [
-            {
-              id: 'test-id',
-              attributes: {
-                timestamp: 1640995199999,
-                status: 'info',
-                message: 'Test log',
-                service: 'test-service',
-              },
-              type: 'log',
-            },
-          ],
-          meta: { page: {} },
-        })
-      })
-
-      const server = setupServer(mockHandler)
-
-      await server.boundary(async () => {
-        const request = createMockToolRequest('get_all_services', {
-          query: '*',
-          from: 1640995100,
-          to: 1640995200,
-          limit: 100,
-        })
-        await toolHandlers.get_all_services(request)
-
-        // Verify storage_tier was included in the filter (API client serializes storageTier as storage_tier)
-        expect(capturedRequestBody).toBeDefined()
-        expect(capturedRequestBody.filter).toBeDefined()
-        expect(capturedRequestBody.filter.storage_tier).toBe('indexes')
-      })()
-
-      server.close()
-    })
-
-    it('should not include storageTier when DATADOG_STORAGE_TIER is not set', async () => {
-      delete process.env.DATADOG_STORAGE_TIER
-
-      let capturedRequestBody: { filter?: { storage_tier?: string } } | null =
-        null
-      const mockHandler = http.post(logsEndpoint, async ({ request }) => {
-        capturedRequestBody = await request.json()
-        return HttpResponse.json({
-          data: [],
-          meta: { page: {} },
-        })
-      })
-
-      const server = setupServer(mockHandler)
-
-      await server.boundary(async () => {
-        const request = createMockToolRequest('get_logs', {
-          query: 'service:test',
-          from: 1640995100,
-          to: 1640995200,
-        })
-        await toolHandlers.get_logs(request)
-
-        // Verify storage_tier was NOT included in the filter
-        expect(capturedRequestBody).toBeDefined()
-        expect(capturedRequestBody.filter).toBeDefined()
-        expect(capturedRequestBody.filter.storage_tier).toBeUndefined()
-      })()
-
-      server.close()
-    })
-
-    it('should handle invalid DATADOG_STORAGE_TIER value gracefully', async () => {
-      process.env.DATADOG_STORAGE_TIER = 'invalid-tier'
-
-      let capturedRequestBody: { filter?: { storage_tier?: string } } | null =
-        null
-      const mockHandler = http.post(logsEndpoint, async ({ request }) => {
-        capturedRequestBody = await request.json()
-        return HttpResponse.json({
-          data: [],
-          meta: { page: {} },
-        })
-      })
-
-      const server = setupServer(mockHandler)
-
-      await server.boundary(async () => {
-        const request = createMockToolRequest('get_logs', {
-          query: 'service:test',
-          from: 1640995100,
-          to: 1640995200,
-        })
-        await toolHandlers.get_logs(request)
-
-        // Verify storage_tier was NOT included when invalid value provided
-        expect(capturedRequestBody).toBeDefined()
-        expect(capturedRequestBody.filter).toBeDefined()
-        expect(capturedRequestBody.filter.storage_tier).toBeUndefined()
-      })()
-
-      server.close()
-    })
-
-    it('should handle case-insensitive DATADOG_STORAGE_TIER values', async () => {
-      process.env.DATADOG_STORAGE_TIER = 'FLEX'
-
-      let capturedRequestBody: { filter?: { storage_tier?: string } } | null =
-        null
-      const mockHandler = http.post(logsEndpoint, async ({ request }) => {
-        capturedRequestBody = await request.json()
-        return HttpResponse.json({
-          data: [],
-          meta: { page: {} },
-        })
-      })
-
-      const server = setupServer(mockHandler)
-
-      await server.boundary(async () => {
-        const request = createMockToolRequest('get_logs', {
-          query: 'service:test',
-          from: 1640995100,
-          to: 1640995200,
-        })
-        await toolHandlers.get_logs(request)
-
-        // Verify storage_tier was normalized to lowercase
-        expect(capturedRequestBody).toBeDefined()
-        expect(capturedRequestBody.filter).toBeDefined()
-        expect(capturedRequestBody.filter.storage_tier).toBe('flex')
-      })()
-
-      server.close()
-    })
-
-    it('should support online-archives storage tier', async () => {
-      process.env.DATADOG_STORAGE_TIER = 'online-archives'
-
-      let capturedRequestBody: { filter?: { storage_tier?: string } } | null =
-        null
-      const mockHandler = http.post(logsEndpoint, async ({ request }) => {
-        capturedRequestBody = await request.json()
-        return HttpResponse.json({
-          data: [],
-          meta: { page: {} },
-        })
-      })
-
-      const server = setupServer(mockHandler)
-
-      await server.boundary(async () => {
-        const request = createMockToolRequest('get_logs', {
-          query: 'service:test',
-          from: 1640995100,
-          to: 1640995200,
-        })
-        await toolHandlers.get_logs(request)
-
-        // Verify online-archives tier is supported
-        expect(capturedRequestBody).toBeDefined()
-        expect(capturedRequestBody.filter).toBeDefined()
-        expect(capturedRequestBody.filter.storage_tier).toBe('online-archives')
       })()
 
       server.close()
